+++ pyanalyze/name_check_visitor.py
@@ -997,7 +997,8 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
     options: Options
     arg_spec_cache: ArgSpecCache
     reexport_tracker: ImplicitReexportTracker
-    being_assigned: Value
+    being_assigned: Optional[Value]
+    is_in_annotated_assignment: False
     current_class: Optional[type]
     current_function_name: Optional[str]
     current_enum_members: Optional[Dict[object, str]]
@@ -1038,6 +1039,7 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
         self.state = VisitorState.collect_names
         # value currently being assigned
         self.being_assigned = AnyValue(AnySource.inference)
+        self.is_in_annotated_assignment = False
         # current match target
         self.match_subject = Composite(AnyValue(AnySource.inference))
         # current class (for inferring the type of cls and self arguments)
@@ -1336,12 +1338,38 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
     def _set_name_in_scope(
         self,
         varname: str,
-        node: object,
+        node: ast.AST,
         value: Value = AnyValue(AnySource.inference),
         *,
         private: bool = False,
+        lookup_node: object = None,
     ) -> Tuple[Value, VarnameOrigin]:
+        if lookup_node is None:
+            lookup_node = node
+
         current_scope = self.scopes.current_scope()
+        if not self.is_in_annotated_assignment and current_scope.is_final(varname):
+            self._show_error_if_checking(
+                node,
+                f"Cannot assign to final name {varname}",
+                ErrorCode.incompatible_assignment,
+            )
+        declared_type = current_scope.get_declared_type(varname)
+        if declared_type is not None:
+            can_assign = declared_type.can_assign(value, self)
+            if isinstance(can_assign, CanAssignError):
+                self._show_error_if_checking(
+                    node,
+                    f"Incompatible assignment: expected {declared_type}, got {value}",
+                    error_code=ErrorCode.incompatible_assignment,
+                    detail=can_assign.display(),
+                )
+
+            if self.is_in_annotated_assignment:
+                # We set the declared type on initial assignment, so that the
+                # annotation can be used to adjust pyanalyze's type inference.
+                value = declared_type
+
         scope_type = current_scope.scope_type
         if self.module is not None and scope_type == ScopeType.module_scope:
             if self.module.__name__ is not None and not private:
@@ -1349,12 +1377,12 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
                     self.module.__name__, varname
                 )
             if varname in current_scope:
-                value, _ = current_scope.get_local(varname, node, self.state)
+                value, _ = current_scope.get_local(varname, lookup_node, self.state)
                 return value, EMPTY_ORIGIN
-        if scope_type == ScopeType.class_scope and isinstance(node, ast.AST):
+        if scope_type == ScopeType.class_scope:
             self._check_for_incompatible_overrides(varname, node, value)
             self._check_for_class_variable_redefinition(varname, node)
-        origin = current_scope.set(varname, value, node, self.state)
+        origin = current_scope.set(varname, value, lookup_node, self.state)
         return value, origin
 
     def _check_for_incompatible_overrides(
@@ -2266,7 +2294,11 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
             ):
                 continue
             self._set_name_in_scope(
-                name, (node, name), KnownValue(value), private=not force_public
+                name,
+                node,
+                KnownValue(value),
+                private=not force_public,
+                lookup_node=(node, name),
             )
 
     def _handle_imports(
@@ -2620,6 +2652,9 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
                 else:
                     target_length += 1
 
+            assert (
+                self.being_assigned is not None
+            ), "annotated assignment can only have a single target"
             being_assigned = unpack_values(
                 self.being_assigned, self, target_length, post_starred_length
             )
@@ -3749,40 +3784,48 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
         if isinstance(annotation, KnownValue) and is_typing_name(
             annotation.val, "Final"
         ):
-            # TODO disallow assignments to Final variables (current code
-            # just avoids false positive errors).
             is_final = True
-            expected_type = AnyValue(AnySource.marker)
+            expected_type = None
         else:
             expected_type = self._value_of_annotation_type(
                 annotation,
                 node.annotation,
                 is_typeddict=self.is_in_typeddict_definition(),
             )
+            # TODO: Also extract Final from more complex annotations
             is_final = False
+
+        if isinstance(node.target, ast.Name):
+            if (
+                not self.scopes.current_scope().set_declared_type(
+                    node.target.id, expected_type, is_final
+                )
+                and self._is_collecting()
+            ):
+                self.show_error(
+                    node,
+                    f"{node.target.id} is already declared",
+                    error_code=ErrorCode.already_declared,
+                )
+        elif isinstance(node.target, ast.Attribute):
+            # TODO: Set declared class attributes.
+            pass
+        else:
+            self.show_error(node, error_code=ErrorCode.invalid_annotated_assignment)
 
         if node.value:
             is_yield = isinstance(node.value, ast.Yield)
             value = self.visit(node.value)
-            can_assign = expected_type.can_assign(value, self)
-            if isinstance(can_assign, CanAssignError):
-                self._show_error_if_checking(
-                    node.value,
-                    f"Incompatible assignment: expected {expected_type}, got {value}",
-                    error_code=ErrorCode.incompatible_assignment,
-                    detail=can_assign.display(),
-                )
-
-            with qcore.override(
-                self, "being_assigned", value if is_final else expected_type
-            ), self.yield_checker.check_yield_result_assignment(is_yield):
-                self.visit(node.target)
         else:
-            with qcore.override(self, "being_assigned", expected_type):
-                self.visit(node.target)
-        # TODO: Idea for what to do if there is no value:
-        # - Scopes keep track of a map {name: expected type}
-        # - Assignments that are inconsistent with the declared type produce an error.
+            is_yield = False
+            value = None
+
+        with qcore.override(
+            self, "being_assigned", value
+        ), self.yield_checker.check_yield_result_assignment(is_yield), qcore.override(
+            self, "is_in_annotated_assignment", True
+        ):
+            self.visit(node.target)
 
     def visit_AugAssign(self, node: ast.AugAssign) -> None:
         is_yield = isinstance(node.value, ast.Yield)
@@ -3822,18 +3865,14 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
                     (ast.stmt, ast.comprehension)
                 )
                 self._name_node_to_statement[node] = statement
-                # If we're in an AnnAssign without a value, we skip the assignment,
-                # since no value is actually assigned to the name.
-                is_ann_assign = (
-                    isinstance(statement, ast.AnnAssign) and statement.value is None
-                )
-            else:
-                is_ann_assign = False
+
             value = self.being_assigned
             origin = EMPTY_ORIGIN
-            if not is_ann_assign:
+            if value is not None:
                 self.yield_checker.record_assignment(node.id)
                 value, origin = self._set_name_in_scope(node.id, node, value=value)
+            else:
+                value = AnyValue(AnySource.inference)
             varname = VarnameWithOrigin(node.id, origin)
             constraint = Constraint(varname, ConstraintType.is_truthy, True, None)
             value = annotate_with_constraint(value, constraint)
@@ -3849,6 +3888,8 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
 
     def visit_arg(self, node: ast.arg) -> None:
         self.yield_checker.record_assignment(node.arg)
+        # it's none only for AnnAssign nodes without a value
+        assert self.being_assigned is not None, "should not happen"
         self._set_name_in_scope(node.arg, node, value=self.being_assigned)
 
     def _should_use_varname_value(self, value: Value) -> bool:
@@ -3906,6 +3947,9 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
         index = index_composite.value
 
         if isinstance(node.ctx, ast.Store):
+            # Happens for AnnAssign without a value. We should have errored already.
+            if self.being_assigned is None:
+                return AnyValue(AnySource.inference)
             if (
                 composite_var is not None
                 and self.scopes.scope_type() == ScopeType.function_scope
@@ -4101,6 +4145,9 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
         root_composite = self.composite_from_node(node.value)
         composite = self._extend_composite(root_composite, node.attr, node)
         if self._is_write_ctx(node.ctx):
+            # AnnAssign without a value
+            if self.being_assigned is None:
+                return Composite(AnyValue(AnySource.inference), composite, node)
             if (
                 composite is not None
                 and self.scopes.scope_type() == ScopeType.function_scope